feat(websocket): add `Response` method for `Message` struct
The `Response` method creates a new `Message` that uses the same event
from the initial `Message` but with the `Data` depending on the value
that is passed as an argument.

This makes it less verbose when WebSocket handlers return a response,
since it only needs to call a method, rather than returning an entire
`Message` struct.